refactor: extract cli/commands/create-app-container/, move config
